+++ pkg/analyzer/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 6.5.0-dev
 * Deprecated `LibraryElement.toLegacyTypeIfOptOut`.
+* Deprecated `LibraryElement.toLegacyElementIfOptOut`.
 
 ## 6.4.1
 * Patch for crash in ffi_verifier.